data: Add new caps to Quassel and Quasseldroid
Add the following capabilities to Quassel and Quasseldroid:
* 'account-tag'
https://github.com/quassel/quassel/commit/d778861aa194578cd90e019b64fe285ea58746c4

* 'echo-message'
https://github.com/quassel/quassel/commit/48017b680ede0dbfb121d1184dfbd13536cfc53f
Disabled by default until batch and labeled-response are supported.
Enable in the meantime with "/CAP REQ echo-message".

* 'invite-notify'
https://github.com/quassel/quassel/commit/e14649614fbbf9b386505a5d782b88b1ac313c1f

* 'message-tags'
https://github.com/quassel/quassel/commit/30269231603184696da66adf4503440062df8962

* 'setname'
https://github.com/quassel/quassel/commit/4e40c486dea949244b73beaf73d5ceb1ef591b5b

The new capabilities are purely core-side, no client changes required.

For the full tests performed, see...
https://github.com/quassel/quassel/pull/545

Clarify SASL EXTERNAL for Quasseldroid - the Android client merely
lacks the configuration UI.  It can connect to a Quassel core with
a certificate associated for SASL EXTERNAL.