Remove freenode from the IRCv3 networks list.
Freenode has recently migrated to a new IRCd so this is no longer
valid. Additionally, consent for it to be listed here was provided
by the previous staff team not the the current staff team. Freenode
staff now also includes someone who is banned from the IRCv3 project
for abusing contributors so it feels inappropriate to include it here.