support lists: Remove vertical overflow+scrollbar + add footer
The overflow+scrollbar makes tables look weird, and I find them hard to use.

This is because vertical scrollbar within a webpage are not very
natural, but we needed them for headers to remain in sight.

With the header repeated at the bottom, this is no longer needed.
Not to mention the header is sticky when scrolling down the page anyway.